Moved solph-related cost calculation into oemof.solph.economics.
Basic LCOE calculation is made in oemof.tools.economics.
Adapted test for LCOE.
